Added Auth ClientService

+++ backend/erp/src/main/java/com/coraduarte/erp/services/ClienteService.java
@@ -1,13 +1,18 @@
 package com.coraduarte.erp.services;
 
+import java.util.Objects;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authorization.AuthorizationDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.coraduarte.erp.models.Cliente;
+import com.coraduarte.erp.models.enums.ProfileEnum;
 import com.coraduarte.erp.repositories.ClienteRepository;
+import com.coraduarte.erp.security.UserSpringSecurity;
+
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -22,23 +27,34 @@ import lombok.Setter;
 @EqualsAndHashCode
 
 public class ClienteService {
-   
+
     @Autowired
     private ClienteRepository clienteRepository;
 
+    public Cliente findById(Long id) {
+        UserSpringSecurity userSpringSecurity = UserService.authenticated();
 
-    public Cliente findById(Long id){
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new AuthorizationDeniedException("Acesso negado!");
+        }
+
         Optional<Cliente> cliente = this.clienteRepository.findById(id);
-        return cliente.orElseThrow(( )-> new RuntimeException(
-            "Usuráio não encontrado! Id: " + id + ", Tipo:" + Cliente.class.getName()
+        return cliente.orElseThrow(() -> new RuntimeException(
+                "Usuráio não encontrado! Id: " + id + ", Tipo:" + Cliente.class.getName()
         ));
     }
 
     @Transactional
     public Cliente create(Cliente obj) {
-      obj.setId(null);
-      obj = this.clienteRepository.save(obj);
-      return obj;
+        UserSpringSecurity userSpringSecurity = UserService.authenticated();
+
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new AuthorizationDeniedException("Acesso negado!");
+        }
+
+        obj.setId(null);
+        obj = this.clienteRepository.save(obj);
+        return obj;
     }
 
     /*@Transactional
@@ -47,15 +63,19 @@ public class ClienteService {
         newObj.setPassword(obj.getPassword());
         return this.clienteRepository.save(newObj);
     } */
+    public void delete(Long id) {
+        UserSpringSecurity userSpringSecurity = UserService.authenticated();
 
-    public void delete(Long id){
+        if (Objects.isNull(userSpringSecurity) || !(userSpringSecurity.hasRole(ProfileEnum.ADMIN))) {
+            throw new AuthorizationDeniedException("Acesso negado!");
+        }
+
         findById(id);
         try {
             this.clienteRepository.deleteById(id);
         } catch (Exception e) {
-           throw new RuntimeException("Não é possível excluir pois há entidades relacionadas!" );
+            throw new RuntimeException("Não é possível excluir pois há entidades relacionadas!");
         }
     }
-
 
 }
